+++ bin/seeder/input/src/configurations/db/db.ejs.sql
@@ -1,0 +1,4 @@
+<%# Quitar los otros # para armar la base de datos vía plantilla y módulos %>
+<%- include("/db/modules/autorizacion.ejs.sql"); %>
+<%- include("/db/modules/ficheros.ejs.sql"); %>
+<%#- include("/db/modules/asamblea.ejs.sql"); %>