+++ bin/seeder/seeder.sh
@@ -1,0 +1,2 @@
+rm -rf output/*
+restero generar --directorio input --salida output